backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
websockets==15.0.1
python-multipart==0.0.20
pydantic==2.11.3
pydantic-settings==2.9.1
faster-whisper==1.1.1
kokoro>=0.9.0
soundfile==0.13.1
numpy==2.2.6
httpx==0.28.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
aiofiles==24.1.0
slowapi>=0.1.9
anthropic>=0.39.0
aiosqlite>=0.19.0
playwright>=1.58.0
# Phase 1.1 — Google Cloud voice provider (optional; only needed when
# VOICE_PROVIDER=google). Local mode works without these installed.
google-cloud-speech>=2.27
google-cloud-texttospeech>=2.16
